+++ stroom-core-client-widget/src/main/java/stroom/widget/tab/client/view/LayerContainerImpl.java
@@ -40,6 +40,7 @@ public class LayerContainerImpl extends Composite implements LayerContainer, Req
     private boolean fade;
     private TransitionTimer transitionTimer;
     private Layer selectedLayer;
+
     public LayerContainerImpl() {
         if (resources == null) {
             resources = GWT.create(Resources.class);